refactor(assistance): MV kb category model to its own file
ref: #783
