+++ app/assistance/serializers/knowledge_base_category.py
@@ -11,7 +11,7 @@ from centurion.serializers.user import UserBaseSerializer
 
 from api.serializers import common
 
-from assistance.models.knowledge_base import KnowledgeBaseCategory
+from assistance.models.knowledge_base_category import KnowledgeBaseCategory
